allow for floating point weight of coffee

+++ app/src/main/java/com/omelan/cofi/components/RecipeInfo.kt
@@ -26,15 +26,15 @@ import com.omelan.cofi.utils.toStringDuration
 import kotlin.math.roundToInt
 
 private data class CoffeeWaterTime(
-    val coffeeWeight: Int = 0,
-    val waterWeight: Int = 0,
+    val coffeeWeight: Double = 0.0,
+    val waterWeight: Double = 0.0,
     val duration: Int = 0,
 ) {
-    fun addCoffee(weight: Int?, time: Int?): CoffeeWaterTime =
-        this.copy(coffeeWeight = coffeeWeight + (weight ?: 0), duration = duration + (time ?: 0))
+    fun addCoffee(weight: Double?, time: Int?): CoffeeWaterTime =
+        this.copy(coffeeWeight = coffeeWeight + (weight ?: 0.0), duration = duration + (time ?: 0))
 
-    fun addWater(weight: Int?, time: Int?): CoffeeWaterTime =
-        this.copy(waterWeight = waterWeight + (weight ?: 0), duration = duration + (time ?: 0))
+    fun addWater(weight: Double?, time: Int?): CoffeeWaterTime =
+        this.copy(waterWeight = waterWeight + (weight ?: 0.0), duration = duration + (time ?: 0))
 
     fun addTime(time: Int?): CoffeeWaterTime = this.copy(duration = duration + (time ?: 0))
 }
@@ -175,13 +175,13 @@ fun RecipeInfoPreview() {
         steps = listOf(
             Step(
                 name = stringResource(R.string.prepopulate_step_coffee),
-                value = 30,
+                value = 30.0,
                 time = 5.toMillis(),
                 type = StepType.ADD_COFFEE,
             ),
             Step(
                 name = stringResource(R.string.prepopulate_step_water),
-                value = 60,
+                value = 60.0,
                 time = 5.toMillis(),
                 type = StepType.WATER,
             ),
@@ -199,13 +199,13 @@ fun RecipeInfoPreview() {
                 name = stringResource(R.string.prepopulate_step_water),
                 time = 30.toMillis(),
                 type = StepType.WATER,
-                value = 240,
+                value = 240.0,
             ),
             Step(
                 name = stringResource(R.string.prepopulate_step_water),
                 time = 30.toMillis(),
                 type = StepType.WATER,
-                value = 200,
+                value = 200.0,
             ),
             Step(
                 name = stringResource(R.string.prepopulate_step_swirl),
@@ -224,13 +224,13 @@ fun RecipeInfoCompactPreview() {
         steps = listOf(
             Step(
                 name = stringResource(R.string.prepopulate_step_coffee),
-                value = 30,
+                value = 30.0,
                 time = 5.toMillis(),
                 type = StepType.ADD_COFFEE,
             ),
             Step(
                 name = stringResource(R.string.prepopulate_step_water),
-                value = 60,
+                value = 60.0,
                 time = 5.toMillis(),
                 type = StepType.WATER,
             ),
@@ -248,13 +248,13 @@ fun RecipeInfoCompactPreview() {
                 name = stringResource(R.string.prepopulate_step_water),
                 time = 30.toMillis(),
                 type = StepType.WATER,
-                value = 240,
+                value = 240.0,
             ),
             Step(
                 name = stringResource(R.string.prepopulate_step_water),
                 time = 30.toMillis(),
                 type = StepType.WATER,
-                value = 200,
+                value = 200.0,
             ),
             Step(
                 name = stringResource(R.string.prepopulate_step_swirl),
